requirements.txt
flask
flask-sqlalchemy
spotipy
yt-dlp
python-dotenv
mutagen
tqdm
colorama
psutil
pillow
requests
gunicorn
